+++ type/collapsedtext/classes/content_type.php
@@ -126,10 +126,12 @@ class content_type extends \mod_unilabel\content_type {
             $useanimation = $this->get_useanimation($unilabel);
 
             $content = [
-                'title'        => $this->get_title($unilabel),
-                'content'      => $intro,
-                'cmid'         => $cm->id,
-                'useanimation' => $useanimation,
+                'title'            => $this->get_title($unilabel),
+                'content'          => $intro,
+                'cmid'             => $cm->id,
+                'useanimation'     => $useanimation,
+                'srtitle_expand'   => get_string('expand'),
+                'srtitle_collapse' => get_string('collapse'),
             ];
 
             if ($cm->id == $cmidfromurl) {