+++ type/courseteaser/classes/content_type.php
@@ -233,9 +233,10 @@ class content_type extends \mod_unilabel\content_type {
                 case 'carousel':
                     $template = 'carousel';
                     if (!empty($this->config->custombutton)) {
+                        $fontbuttons = setting_configselect_button::get_font_buttons();
                         $content['custombuttons']   = 1;
-                        $content['fontawesomenext'] = setting_configselect_button::$buttonlist[$this->config->custombutton]['next'];
-                        $content['fontawesomeprev'] = setting_configselect_button::$buttonlist[$this->config->custombutton]['prev'];
+                        $content['fontawesomenext'] = $fontbuttons[$this->config->custombutton]['next'];
+                        $content['fontawesomeprev'] = $fontbuttons[$this->config->custombutton]['prev'];
 
                         // To make sure we have clean html we have to put the carousel css into the <head> by using javascript.
                         $cssstring                = $renderer->render_from_template('mod_unilabel/carousel_button_style', $content);